refactor: split components into multiple modules
Also improve tests and use protocol for macro events
